requirements.txt
streamlit
pandas
crawl4ai
pydantic
python-dotenv
httpx
chromadb>=0.5.0,<0.7
pytest
